Test config and display more errors
The two threads for handling messages and updating the command now do a
health check before starting operations, and more of the errors are
displayed in the GUI. Additionally, move welcome counter `let` statement
out of loop. This would cause every welcome message to be treated as the
first welcome message, re-triggering the subscription logic. If Twitch
sends a Reconnect message, the reconnection logic involved receiving
another Welcome message. Sine the subscription persists, it is not
necessary to repeat any of that logic.